// cmd/client-app/nextjs/src/proto/generated/ElementRule.ts
// Original file: ../mediatext.proto


export interface ElementRule {
  'start'?: (number);
  'stop'?: (number);
  'name'?: (string);
  'note'?: (string);
  'color'?: (string);
}

export interface ElementRule__Output {
  'start': (number);
  'stop': (number);
  'name': (string);
  'note': (string);
  'color': (string);
}